+++ test/ash_scylla/dsl_repo_migration_test.exs
@@ -347,7 +347,7 @@ defmodule AshScylla.DslRepoMigrationContinuedTest do
   describe "AshScylla.version/0" do
     test "returns the current version string" do
       version = AshScylla.version()
-      assert version == "0.3.0"
+      assert version == "0.4.0"
     end
 
     test "returns a string" do